requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.6
firecrawl-py==1.11.1
frozenlist==1.5.0
gotrue==2.11.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith==0.2.10
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.1
openai==1.59.5
orjson==3.10.14
packaging==24.2
playwright==1.49.1
postgrest==0.19.1
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.1.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==13.1
wheel==0.44.0
yarl==1.18.3
